+++ app/src/app/admin/sales/sale-returns/sale-returns.component.ts
@@ -1891,7 +1891,7 @@ async autoFillProductDetails(field, data) {
                         const quantity = field.form.controls.quantity.value;
                         const rate = data;
                         if (rate && quantity) {
-                          field.form.controls.amount.setValue(parseInt(rate) * parseInt(quantity));
+                          field.form.controls.amount.setValue(parseFloat(rate) * parseFloat(quantity));
                         }
                       }
                     });
