+++ app/src/app/admin/sales/sales.component.ts
@@ -7775,7 +7775,7 @@ createChildOrdersForProducts(productDetails, saleOrderDetails, orderAttachments,
                         const quantity = field.form.controls.quantity.value;
                         const rate = data;
                         if (rate && quantity) {
-                          field.form.controls.amount.setValue(parseInt(rate) * parseInt(quantity));
+                          field.form.controls.amount.setValue(parseFloat(rate) * parseFloat(quantity));
                         }
                       }
                       this.totalAmountCal();
