+++ app/src/app/admin/sales/salesinvoice/salesinvoice.component.ts
@@ -2804,7 +2804,7 @@ createSaleInovice() {
                         const quantity = field.form.controls.quantity.value;
                         const rate = data;
                         if (rate && quantity) {
-                          field.form.controls.amount.setValue(parseInt(rate) * parseInt(quantity));
+                          field.form.controls.amount.setValue(parseFloat(rate) * parseFloat(quantity));
                         }
                       }
                       this.totalAmountCal();
